Update CleanRip.txt
Added information that the app can be used in the vWii. (Wii U)
Added information that the app can be used to backup GameCube games on the original Wii only.
What the heck is a "tarjeta USB" (USB Card)? This is changed to "tarjeta SD o unidad USB" (SD Card or a USB drive)
Added a notice that the backup can be used on backup loaders and provided 3 examples.