WiiLink24 Japanese Version
Wow! This was a fun task.
As WiiLink24 concerns Japanese channels specifically, I thought that it would be a good idea to create a Japanese translation for the installation instructions page. This was not created with a translator, and done entirely by me! It is possible that there are sections that could be improved as I am not a native Japanese person, however my translation is certainly understandable and more clear than a machine translation of the English page.
If you have any questions, please contact me on Discord: `Noren#7588`.