Notes about region selection and SD adapters
Ever since GeoIP broke for the Letterbomb site, 4.3E has been the region selected by default, and users have been thrown off by this. So I added an extra note explaining this and confirming that the user has chosen the proper region. Also added a note to step 6 about what could happen if the incorrect region is selected.

Believe it or not, some users have had issues with Letterbomb because they thought a USB-SD adapter plugged into the Wii would still work. So added a note about that to step 4.