Revert LetterBomb page
I don't know why Naim2000 made some kind of change to the page. Techchrism.me is nicer since it doesn't have the confusing "Cut the red wire" and the "Cut the blue wire" options, the page looks nicer, and it's web-based. It has a dark background as well, which isn't important, the only problem I have is that the cooler pride logo isn't on the LetterBomb page.