Replace WiiLink guide with their website
It's been nearly 3 weeks since there had been drama. In that time, a lot of misunderstandings have been made as to what I had been trying to do with our guide. We as a group have decided to just link to their website for the installation, something that I was trying to do in first place but it got soured by the drama.